(ASC-233) Verify galera cluster survived the physical device reboots
This PR adds a test of verifying galera cluster is running properly after the physical device reboots

Also, the delay after the last reboot is changed to 5 min. (from 2 minutes prior to this PR) because 2 min. delay causes some Nova tests to fail, looks like the default availability zone is not available two minutes after nova reboot.
